Module bacon_sci_1::constants

Constants

G

Newtonian gravitational constant, G, m^3 / kg s^2

G_uncertainty

R

Molar gas constant, R, J / mol K

avogadro

Avogadro constant, N_A, 1/ mol

boltzmann

Boltzmann constant, k, J/K

c

Speed of light in a vacuum, m/s

e_charge

Elementary charge, e, C

electron_mass

Mass of an electron, kg

electron_mass_uncertainty

fine_structure

Fine structure constant, α, unitless

fine_structure_uncertainty

g

Gravitational acceleration on Earth’s surface, m / s^2

h

Plank constant, h, J

h_bar

Reduced plank constant, ħ, J s

neutron_mass

Mass of an neutron kg

neutron_mass_uncertainty

permeability

Permeability of free space, μ₀, N/A^2

permeability_uncertainty

permittivity

Permittivity of free space, ε₀, F/

permittivity_uncertainty

proton_mass

Mass of an proton kg

proton_mass_uncertainty

rydberg

Rydberg constant, R_inf, 1/m

rydberg_uncertainty

stefan_boltzmann

Stefan-Boltzmann constant, σ, W / m^2 K^4

wien

Wien displacement law constant, b, wavelength, m K

wien_frequency

Wien displacement law constant, b’, frequency, Hz /K

Statics

CODATA

NIST CODATA. Maps a string of a value name to a triplet containing an f64 of its value, an f64 of its uncertainty, and a string of its units