Trait bacon_sci_1::ivp::IVPSolver

pub trait IVPSolver<N: ComplexField, S: DimName>: Sized where
    DefaultAllocator: Allocator<N, S>, {
    fn step<T: Clone, F: FnMut(N::RealField, &[N], &mut T) -> Result<VectorN<N, S>, String>>(
        &mut self, 
        f: F, 
        params: &mut T
    ) -> Result<IVPStatus<N, S>, String>;
    fn with_tolerance(self, tol: N::RealField) -> Result<Self, String>;
    fn with_dt_max(self, max: N::RealField) -> Result<Self, String>;
    fn with_dt_min(self, min: N::RealField) -> Result<Self, String>;
    fn with_start(self, t_initial: N::RealField) -> Result<Self, String>;
    fn with_end(self, t_final: N::RealField) -> Result<Self, String>;
    fn with_initial_conditions(self, start: &[N]) -> Result<Self, String>;
    fn build(self) -> Self;
    fn get_initial_conditions(&self) -> Option<VectorN<N, S>>;
    fn get_time(&self) -> Option<N::RealField>;
    fn check_start(&self) -> Result<(), String>;

    fn solve_ivp<T: Clone, F: FnMut(N::RealField, &[N], &mut T) -> Result<VectorN<N, S>, String>>(
        self, 
        f: F, 
        params: &mut T
    ) -> Result<Vec<(N::RealField, VectorN<N, S>)>, String> { ... }
}

Trait defining what it means to be an IVP solver. solve_ivp is automatically implemented based on your step implementation.

Required Methods

fn step<T: Clone, F: FnMut(N::RealField, &[N], &mut T) -> Result<VectorN<N, S>, String>>(
&mut self,
f: F,
params: &mut T
) -> Result<IVPStatus<N, S>, String>

Step forward in the solver. Returns if the solver is finished, produced an acceptable state, or needs to be redone.

fn with_tolerance(self, tol: N::RealField) -> Result<Self, String>

Set the error tolerance for this solver.

fn with_dt_max(self, max: N::RealField) -> Result<Self, String>

Set the maximum time step for this solver.

fn with_dt_min(self, min: N::RealField) -> Result<Self, String>

Set the minimum time step for this solver.

fn with_start(self, t_initial: N::RealField) -> Result<Self, String>

Set the initial time for this solver.

fn with_end(self, t_final: N::RealField) -> Result<Self, String>

Set the end time for this solver.

fn with_initial_conditions(self, start: &[N]) -> Result<Self, String>

Set the initial conditions for this solver.

fn build(self) -> Self

Build this solver.

fn get_initial_conditions(&self) -> Option<VectorN<N, S>>

Return the initial conditions. Called once at the very start of solving.

fn get_time(&self) -> Option<N::RealField>

Get the current time of the solver.

fn check_start(&self) -> Result<(), String>

Make sure that every value that needs to be set is set before the solver starts

Provided Methods

fn solve_ivp<T: Clone, F: FnMut(N::RealField, &[N], &mut T) -> Result<VectorN<N, S>, String>>(
self,
f: F,
params: &mut T
) -> Result<Vec<(N::RealField, VectorN<N, S>)>, String>

Solve an initial value problem, consuming the solver