Struct bacon_sci_1::polynomial::Polynomial

pub struct Polynomial<N: ComplexField> { /* private fields */ }

Polynomial on a ComplexField.

Implementations

impl<N: ComplexField> Polynomial<N>

pub fn new() -> Self

Returns the zero polynomial on a given field

pub fn with_tolerance(
tolerance: <N as ComplexField>::RealField
) -> Result<Self, String>

pub fn with_capacity(capacity: usize) -> Self

Returns the zero polynomial on a given field with preallocated memory

pub fn from_slice(data: &[N]) -> Self

Create a polynomial from a slice, with the first element of the slice being the highest power

pub fn set_tolerance(
&mut self,
tolerance: <N as ComplexField>::RealField
) -> Result<(), String>

pub fn get_tolerance(&self) -> <N as ComplexField>::RealField

pub fn order(&self) -> usize

Get the order of the polynomial

pub fn get_coefficients(&self) -> Vec<N>

Returns the coefficients in the correct order to recreate the polynomial with Polynomial::from_slice(data: &[N]);

pub fn get_coefficient(&self, ind: usize) -> N

Get the coefficient of a power

pub fn make_complex(
&self
) -> Polynomial<Complex<<N as ComplexField>::RealField>>

Make a polynomial complex

pub fn evaluate(&self, x: N) -> N

Evaluate a polynomial at a value

pub fn evaluate_derivative(&self, x: N) -> (N, N)

Evaluate a polynomial and its derivative at a value

pub fn set_coefficient(&mut self, power: u32, coefficient: N)

Set a coefficient of a power in the polynomial

pub fn purge_coefficient(&mut self, power: usize)

Remove the coefficient of a power in the polynomial

pub fn purge_leading(&mut self)

Remove all leading 0 coefficients

pub fn derivative(&self) -> Self

Get the derivative of the polynomial

pub fn antiderivative(&self, constant: N) -> Self

Get the antiderivative of the polynomial with specified constant

pub fn integrate(&self, lower: N, upper: N) -> N

Integrate this polynomial between to starting points

pub fn divide(&self, divisor: &Polynomial<N>) -> Result<(Self, Self), String>

Divide this polynomial by another, getting a quotient and remainder, using tol to check for 0

pub fn roots(
&self,
tol: <N as ComplexField>::RealField,
n_max: usize
) -> Result<VecDeque<Complex<<N as ComplexField>::RealField>>, String>

Get the n (possibly including repeats) of the polynomial given n using Laguerre’s method

pub fn dft(&self, size: usize) -> Vec<Complex<<N as ComplexField>::RealField>>

Get the polynomial in point form evaluated at roots of unity at k points where k is the smallest power of 2 greater than or equal to size

pub fn idft(
vec: &[Complex<<N as ComplexField>::RealField>],
tol: <N as ComplexField>::RealField
) -> Self